- Added max total file size for file upload. - Grouped files in based on screen size. - Moved SegmentedDatePicker to shared components
