+++ Website/src/components/create/mobile/CreateStepIndicator.tsx
