+++ Website/src/components/create/mobile/MobileUploadArea.tsx
